requirements.txt
# Django
Django==3.0.13
# Django REST Framework
djangorestframework==3.12.1
# ------------------------------------------------------------------------------
django-model-utils==4.0.0
django-simple-history==2.11.0
djangorestframework-simplejwt==4.6.0
drf-rw-serializers==1.0.3
pycodestyle==2.8.0
asgiref==3.5.2
autopep8==1.6.0
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2